backend/requirements.txt
alipay==0.7.4
amqp==2.2.2
asn1crypto==0.24.0
beautifulsoup4==4.7.1
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.2.2
django-cors-headers==2.4.0
django-filter==2.1.0
django-haystack==2.8.1
django-redis==4.8.0
django-rest-swagger==2.2.0
django-simpleui==1.4.8
django-tinymce==2.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
itypes==1.1.0
jieba==0.39
Jinja2==2.10
kombu==4.1.0
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
olefile==0.44
openapi-codec==1.3.2
pefile==2018.8.8
Pillow==5.4.1
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.7.3
pyflakes==2.1.1
PyJWT==1.6.4
python-alipay-sdk==1.7.0
python-ptrace==0.9.3
pytz==2017.2
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.8
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
Whoosh==2.7.4
xlrd==1.2.0
yapf==0.26.0
